requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
beautifulsoup4==4.13.4
    # via -r requirements.in
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
fake-useragent==2.2.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via python-telegram-bot
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
instagrapi==2.1.5
    # via -r requirements.in
pycryptodomex==3.23.0
    # via instagrapi
pydantic==2.11.5
    # via instagrapi
pysocks==1.7.1
    # via instagrapi
python-dotenv==1.1.0
    # via -r requirements.in
python-telegram-bot==22.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   instagrapi
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
typing-extensions==4.14.0
    # via
    #   anyio
    #   beautifulsoup4
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via requests
